pyproject.toml
[project]
name = "cbioportal-article-classification"
version = "0.1.0"
description = "Tool for fetching and classifying papers that cite cBioPortal"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic[bedrock]>=0.39.0",
    "instructor>=1.0.0",
    "pydantic>=2.0.0",
    "biopython>=1.83",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pandas>=2.1.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "python-dotenv>=1.0.0",
    "pypdf2>=3.0.0",
    "tqdm>=4.66.0",
    "click>=8.1.0",
]

[project.scripts]
cbioportal-classify = "cbioportal_article_classification.main:cli"

[tool.uv]
package = true
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
